Add files via upload
This is a basic CRM system designed for two categories: User and Manager. Each has its subcategories and features and both inherit the Person class, meaning every feature of the Person class, may be used on the User and Manager class. The users of the system have a unique ID which increments for every user registered. The program presents a menu for the simple user in which he can login by writing the username and password. In the menu there are features like: new service (remove resolve), new costumer, different offers and some of them are accessible only by the managers like: Overall number of Services requested by customers and Total amount of all fees of resolved Services are found.
Each user is able to change from manager to user and vice versa. With the usage of inheritance in the classes it is able to access more attributes easily. For example there are three Classes inheriting from Customer class: Simple Customer, Loyal Customer, and Business Customer.
The program has two main classes: One for the console version and the second main class for executing the application with the GUI. Before closing the system, the program will save the data by using serialization, so they can be loaded later.